Moving this test into its own space
